w4-deployment/stream/function/requirements.txt
mlflow
google-cloud-pubsub
scikit-learn==1.0.2
python-dotenv
boto3
